requirements.txt
pafy
python-vlc
youtube-dl==2020.12.2